requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
frozenlist==1.4.1
idna==3.7
joblib==1.4.2
multidict==6.0.5
numpy==1.26.4
pillow==10.3.0
psycopg2-binary==2.9.9
requests==2.32.2
scikit-learn==1.5.0
scipy==1.13.1
sqlparse==0.5.0
threadpoolctl==3.5.0
tqdm==4.66.4
urllib3==2.2.1
yarl==1.9.4
